+++ packages/vexide-devices/src/adi/accelerometer.rs
@@ -127,11 +127,9 @@ impl Sensitivity {
     }
 }
 
-impl AdiDevice for AdiAccelerometer {
-    type PortNumberOutput = u8;
-
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+impl AdiDevice<1> for AdiAccelerometer {
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {