+++ packages/vexide-devices/src/adi/addrled.rs
@@ -149,11 +149,9 @@ impl AdiAddrLed {
     }
 }
 
-impl AdiDevice for AdiAddrLed {
-    type PortNumberOutput = u8;
-
-    fn port_number(&self) -> Self::PortNumberOutput {
-        self.port.number()
+impl AdiDevice<1> for AdiAddrLed {
+    fn port_numbers(&self) -> [u8; 1] {
+        [self.port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {