+++ packages/vexide-devices/src/adi/encoder.rs
@@ -48,7 +48,6 @@
 //! [`RotationSensor`]: crate::smart::rotation::RotationSensor
 //! [`SmartPort`]: crate::smart::SmartPort
 
-use snafu::{ensure, Snafu};
 use vex_sdk::{vexDeviceAdiValueGet, vexDeviceAdiValueSet};
 
 use super::{AdiDevice, AdiDeviceType, AdiPort};
@@ -67,19 +66,11 @@ impl AdiEncoder {
 
     /// Create a new encoder sensor from a top and bottom [`AdiPort`].
     ///
-    /// ```no_run
-    /// # fn make_encoder(peripherals: Peripherals) -> Result<(), EncoderError> {
-    /// let encoder = AdiEncoder::new((peripherals.adi_a, peripherals.adi_b))?;
-    /// # Ok(())
-    /// # }
-    /// ```
+    /// # Panics
     ///
-    /// # Errors
-    ///
-    /// - If the top and bottom ports originate from different [`AdiExpander`](crate::smart::expander::AdiExpander)s,
-    ///   returns [`EncoderError::ExpanderPortMismatch`].
+    /// - If the top and bottom ports originate from different [`AdiExpander`](crate::smart::expander::AdiExpander)s.
     /// - If the ports are not directly next to each other or in an invalid position (one port is not in A, C, E, G and
-    ///   the other is not in in B, D, F), returns [`EncoderError::BadPortPlacement`].
+    ///   the other is not in in B, D, F).
     ///
     /// # Examples
     ///
@@ -92,7 +83,7 @@ impl AdiEncoder {
     ///
     /// #[vexide::main]
     /// async fn main(peripherals: Peripherals) {
-    ///     let encoder = AdiEncoder::new((peripherals.adi_a, peripherals.adi_b)).expect("could not create encoder");
+    ///     let encoder = AdiEncoder::new(peripherals.adi_a, peripherals.adi_b);
     ///
     ///     loop {
     ///         println!("encoder position: {:?}", encoder.position());
@@ -100,7 +91,8 @@ impl AdiEncoder {
     ///     }
     /// }
     /// ```
-    pub fn new((top_port, bottom_port): (AdiPort, AdiPort)) -> Result<Self, EncoderError> {
+    #[must_use]
+    pub fn new(top_port: AdiPort, bottom_port: AdiPort) -> Self {
         let top_number = top_port.number();
         let bottom_number = bottom_port.number();
 
@@ -109,34 +101,32 @@ impl AdiEncoder {
         // Might be fixed through #120
 
         // Top and bottom must be plugged into the same ADI expander.
-        ensure!(
+        assert!(
             top_port.expander_index() == bottom_port.expander_index(),
-            ExpanderPortMismatchSnafu {
-                top_port_expander: top_port.expander_number(),
-                bottom_port_expander: bottom_port.expander_number()
-            }
+            "The specified top and bottom ports belong to different ADI expanders. Both expanders {:?} and {:?} were provided.",
+            top_port.expander_number(),
+            bottom_port.expander_number(),
         );
 
         // Top and bottom must be some combination of (AB, CD, EF, GH) or (BA, CD, FE, HG)
         let top_is_even = top_number % 2 == 0;
-        ensure!(
+        assert!(
             if top_is_even {
                 bottom_number == top_number - 1
             } else {
                 bottom_number == top_number + 1
             },
-            BadBottomPortSnafu {
-                top_port: top_port.number(),
-                bottom_port: bottom_port.number()
-            }
+            "Encoder ports must be placed directly next to each other and in some combination of AB, CD, EF, GH, or BA, CD, EF, HG. (Got `{}{}`)",
+            adi_port_name(top_number),
+            adi_port_name(bottom_number),
         );
 
         top_port.configure(AdiDeviceType::Encoder);
 
-        Ok(Self {
+        Self {
             top_port,
             bottom_port,
-        })
+        }
     }
 
     /// Returns the distance reading of the encoder sensor in centimeters.
@@ -145,7 +135,9 @@ impl AdiEncoder {
     ///
     /// # Errors
     ///
-    /// If the ADI device could not be accessed, returns [`EncoderError::Port`].
+    /// - A [`PortError::Disconnected`] error is returned if an ADI expander device was required but not connected.
+    /// - A [`PortError::IncorrectDevice`] error is returned if an ADI expander device was required but
+    ///   something else was connected.
     ///
     /// # Examples
     ///
@@ -158,7 +150,7 @@ impl AdiEncoder {
     ///
     /// #[vexide::main]
     /// async fn main(peripherals: Peripherals) {
-    ///     let encoder = AdiEncoder::new((peripherals.adi_a, peripherals.adi_b)).expect("could not create encoder");
+    ///     let encoder = AdiEncoder::new(peripherals.adi_a, peripherals.adi_b);
     ///
     ///     loop {
     ///         println!("encoder position: {:?}", encoder.position());
@@ -166,7 +158,7 @@ impl AdiEncoder {
     ///     }
     /// }
     /// ```
-    pub fn position(&self) -> Result<Position, EncoderError> {
+    pub fn position(&self) -> Result<Position, PortError> {
         self.top_port.validate_expander()?;
         self.top_port.configure(self.device_type());
 
@@ -188,7 +180,9 @@ impl AdiEncoder {
     ///
     /// # Errors
     ///
-    /// If the ADI device could not be accessed, returns [`EncoderError::Port`].
+    /// - A [`PortError::Disconnected`] error is returned if an ADI expander device was required but not connected.
+    /// - A [`PortError::IncorrectDevice`] error is returned if an ADI expander device was required but
+    ///   something else was connected.
     ///
     /// # Examples
     ///
@@ -201,13 +195,13 @@ impl AdiEncoder {
     ///
     /// #[vexide::main]
     /// async fn main(peripherals: Peripherals) {
-    ///     let encoder = AdiEncoder::new((peripherals.adi_a, peripherals.adi_b)).expect("could not create encoder");
+    ///     let encoder = AdiEncoder::new(peripherals.adi_a, peripherals.adi_b);
     ///
     ///     // Treat the encoder as if it were at 180 degrees.
     ///     _ = encoder.set_position(Position::from_degrees(180));
     /// }
     /// ```
-    pub fn set_position(&self, position: Position) -> Result<(), EncoderError> {
+    pub fn set_position(&self, position: Position) -> Result<(), PortError> {
         self.top_port.validate_expander()?;
 
         unsafe {
@@ -228,7 +222,9 @@ impl AdiEncoder {
     ///
     /// # Errors
     ///
-    /// If the ADI device could not be accessed, returns [`EncoderError::Port`].
+    /// - A [`PortError::Disconnected`] error is returned if an ADI expander device was required but not connected.
+    /// - A [`PortError::IncorrectDevice`] error is returned if an ADI expander device was required but
+    ///   something else was connected.
     ///
     /// # Examples
     ///
@@ -241,23 +237,21 @@ impl AdiEncoder {
     ///
     /// #[vexide::main]
     /// async fn main(peripherals: Peripherals) {
-    ///     let encoder = AdiEncoder::new((peripherals.adi_a, peripherals.adi_b)).expect("could not create encoder");
+    ///     let encoder = AdiEncoder::new(peripherals.adi_a, peripherals.adi_b);
     ///
     ///     // Reset the encoder position to zero.
     ///     // This doesn't really do anything in this case, but it's a good example.
     ///     _ = encoder.reset_position();
     /// }
     /// ```
-    pub fn reset_position(&mut self) -> Result<(), EncoderError> {
+    pub fn reset_position(&mut self) -> Result<(), PortError> {
         self.set_position(Position::default())
     }
 }
 
-impl AdiDevice for AdiEncoder {
-    type PortNumberOutput = (u8, u8);
-
-    fn port_number(&self) -> Self::PortNumberOutput {
-        (self.top_port.number(), self.bottom_port.number())
+impl AdiDevice<2> for AdiEncoder {
+    fn port_numbers(&self) -> [u8; 2] {
+        [self.top_port.number(), self.bottom_port.number()]
     }
 
     fn expander_port_number(&self) -> Option<u8> {
@@ -267,52 +261,4 @@ impl AdiDevice for AdiEncoder {
     fn device_type(&self) -> AdiDeviceType {
         AdiDeviceType::Encoder
     }
-}
-
-#[derive(Debug, Snafu)]
-/// Errors that can occur when interacting with an encoder range finder.
-pub enum EncoderError {
-    /// The top wire must be on an odd numbered port (A, C, E, G).
-    #[snafu(display(
-        "The top ADI port provided (`{}`) was not odd numbered (A, C, E, G).",
-        adi_port_name(*port)
-    ))]
-    BadTopPort {
-        /// The port number that caused the error.
-        port: u8,
-    },
-
-    /// Ports must be placed directly next to each other, with the ports being some combination of (AB, CD, EF, GH) or (BA, CD, EF, HG).
-    // TODO: Change this to be named `BadPortPlacement` in the next major release.
-    #[snafu(display(
-        "Encoder ports must be placed directly next to each other and in some combination of AB, CD, EF, GH, or BA, CD, EF, HG. (Got `{}{}`)",
-        adi_port_name(*top_port),
-        adi_port_name(*bottom_port),
-    ))]
-    BadBottomPort {
-        /// The bottom port number that caused the error.
-        bottom_port: u8,
-        /// The top port number that caused the error.
-        top_port: u8,
-    },
-
-    /// The specified top and bottom ports may not belong to different ADI expanders.
-    #[snafu(display(
-        "The specified top and bottom ports may not belong to different ADI expanders. Both expanders {:?} and {:?} were provided.",
-        top_port_expander,
-        bottom_port_expander
-    ))]
-    ExpanderPortMismatch {
-        /// The top port's expander number.
-        top_port_expander: Option<u8>,
-        /// The bottom port's expander number.
-        bottom_port_expander: Option<u8>,
-    },
-
-    /// Generic port related error.
-    #[snafu(transparent)]
-    Port {
-        /// The source of the error.
-        source: PortError,
-    },
 }